Moving Cache Exceptions to their own namespace.
