+++ src/Cache/CacheEngine.php
@@ -16,6 +16,7 @@ declare(strict_types=1);
  */
 namespace Cake\Cache;
 
+use Cake\Cache\Exception\InvalidArgumentException;
 use Cake\Core\InstanceConfigTrait;
 use DateInterval;
 use DateTime;
@@ -96,7 +97,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *
      * @param string $key Key to check.
      * @return void
-     * @throws \Cake\Cache\InvalidArgumentException When the key is not valid.
+     * @throws \Cake\Cache\Exception\InvalidArgumentException When the key is not valid.
      */
     protected function ensureValidKey($key): void
     {
@@ -111,7 +112,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      * @param iterable $iterable The iterable to check.
      * @param string $check Whether to check keys or values.
      * @return void
-     * @throws \Cake\Cache\InvalidArgumentException
+     * @throws \Cake\Cache\Exception\InvalidArgumentException
      */
     protected function ensureValidType($iterable, string $check = self::CHECK_VALUE): void
     {
@@ -137,7 +138,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      * @param iterable $keys A list of keys that can obtained in a single operation.
      * @param mixed $default Default value to return for keys that do not exist.
      * @return iterable A list of key value pairs. Cache keys that do not exist or are stale will have $default as value.
-     * @throws \Cake\Cache\InvalidArgumentException If $keys is neither an array nor a Traversable,
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If $keys is neither an array nor a Traversable,
      *   or if any of the $keys are not a legal value.
      */
     public function getMultiple($keys, $default = null): iterable
@@ -160,7 +161,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *   the driver supports TTL then the library may set a default value
      *   for it or let the driver take care of that.
      * @return bool True on success and false on failure.
-     * @throws \Cake\Cache\InvalidArgumentException If $values is neither an array nor a Traversable,
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If $values is neither an array nor a Traversable,
      *   or if any of the $values are not a legal value.
      */
     public function setMultiple($values, $ttl = null): bool
@@ -196,7 +197,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *
      * @param iterable $keys A list of string-based keys to be deleted.
      * @return bool True if the items were successfully removed. False if there was an error.
-     * @throws \Cake\Cache\InvalidArgumentException If $keys is neither an array nor a Traversable,
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If $keys is neither an array nor a Traversable,
      *   or if any of the $keys are not a legal value.
      */
     public function deleteMultiple($keys): bool
@@ -223,7 +224,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *
      * @param string $key The cache item key.
      * @return bool
-     * @throws \Cake\Cache\InvalidArgumentException If the $key string is not a legal value.
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If the $key string is not a legal value.
      */
     public function has($key): bool
     {
@@ -236,7 +237,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      * @param string $key The unique key of this item in the cache.
      * @param mixed $default Default value to return if the key does not exist.
      * @return mixed The value of the item from the cache, or $default in case of cache miss.
-     * @throws \Cake\Cache\InvalidArgumentException If the $key string is not a legal value.
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If the $key string is not a legal value.
      */
     abstract public function get($key, $default = null);
 
@@ -249,7 +250,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *   the driver supports TTL then the library may set a default value
      *   for it or let the driver take care of that.
      * @return bool True on success and false on failure.
-     * @throws \Cake\Cache\InvalidArgumentException
+     * @throws \Cake\Cache\Exception\InvalidArgumentException
      *   MUST be thrown if the $key string is not a legal value.
      */
     abstract public function set($key, $value, $ttl = null): bool;
@@ -337,7 +338,7 @@ abstract class CacheEngine implements CacheInterface, CacheEngineInterface
      *
      * @param string $key the key passed over
      * @return string Prefixed key with potentially unsafe characters replaced.
-     * @throws \Cake\Cache\InvalidArgumentException If key's value is invalid.
+     * @throws \Cake\Cache\Exception\InvalidArgumentException If key's value is invalid.
      */
     protected function _key($key): string
     {
