+++ src/Cache/Engine/MemcachedEngine.php
@@ -16,8 +16,8 @@ declare(strict_types=1);
  */
 namespace Cake\Cache\Engine;
 
+use Cake\Cache\Exception\InvalidArgumentException;
 use Cake\Cache\CacheEngine;
-use InvalidArgumentException;
 use Memcached;
 use RuntimeException;
 
@@ -98,7 +98,7 @@ class MemcachedEngine extends CacheEngine
      *
      * @param array<string, mixed> $config array of setting for the engine
      * @return bool True if the engine has been successfully initialized, false if not
-     * @throws \InvalidArgumentException When you try use authentication without
+     * @throws \Cake\Cache\Exception\InvalidArgumentException When you try use authentication without
      *   Memcached compiled with SASL support
      */
     public function init(array $config = []): bool
@@ -199,7 +199,7 @@ class MemcachedEngine extends CacheEngine
      * Settings the memcached instance
      *
      * @return void
-     * @throws \InvalidArgumentException When the Memcached extension is not built
+     * @throws \Cake\Cache\Exception\InvalidArgumentException When the Memcached extension is not built
      *   with the desired serializer engine.
      */
     protected function _setOptions(): void
