+++ src/Cache/Exception/CacheWriteException.php
@@ -14,14 +14,13 @@ declare(strict_types=1);
  * @since         3.7.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Cache;
+namespace Cake\Cache\Exception;
 
 use Cake\Core\Exception\CakeException;
-use Psr\SimpleCache\InvalidArgumentException as InvalidArgumentInterface;
 
 /**
  * Exception raised when cache keys are invalid.
  */
-class InvalidArgumentException extends CakeException implements InvalidArgumentInterface
+class CacheWriteException extends CakeException
 {
 }
