+++ src/Cache/Exception/InvalidArgumentException.php
@@ -14,13 +14,14 @@ declare(strict_types=1);
  * @since         3.7.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Cache;
+namespace Cake\Cache\Exception;
 
 use Cake\Core\Exception\CakeException;
+use Psr\SimpleCache\InvalidArgumentException as InvalidArgumentInterface;
 
 /**
  * Exception raised when cache keys are invalid.
  */
-class CacheWriteException extends CakeException
+class InvalidArgumentException extends CakeException implements InvalidArgumentInterface
 {
 }
