+++ tests/TestCase/Cache/CacheTest.php
@@ -19,10 +19,10 @@ namespace Cake\Test\TestCase\Cache;
 use BadMethodCallException;
 use Cake\Cache\Cache;
 use Cake\Cache\CacheRegistry;
-use Cake\Cache\CacheWriteException;
+use Cake\Cache\Exception\CacheWriteException;
 use Cake\Cache\Engine\FileEngine;
 use Cake\Cache\Engine\NullEngine;
-use Cake\Cache\InvalidArgumentException;
+use Cake\Cache\Exception\InvalidArgumentException;
 use Cake\TestSuite\TestCase;
 use Psr\SimpleCache\CacheInterface as SimpleCacheInterface;
 use stdClass;
